requirements.txt
Flask==2.0.1
Pillow==8.2.0
tensorflow==2.5.0
tensorflow-hub==0.12.0
